Phase 1/4 (sign-conversion cleanup): TravelingSalesmanSolution::_solution becomes vector<uint32_t>.
_solution holds city indices, which are always non-negative, but was
declared vector<int> while everything that indexes with those values
(Matrix<double,2>::operator[], std::vector<double>::operator[]) wants
an unsigned index -- the dominant source of the 49 -Wsign-conversion
sites found while auditing this codebase with -Weverything. This is
phase 1 of a 4-phase plan (see this session's plan doc) to unify types
at the root rather than paper over the mismatch with casts.

Follow-on changes forced by the member type change: randomize()'s
_solution[i] = static_cast<int>(i) simplifies to a plain assignment;
generateNeighbor()'s holder and switch-index truncation become
uint32_t (the truncated value is always non-negative by construction,
so this isn't a behavior change, just matching the type to the value);
dump()'s ostream_iterator<int> becomes <uint32_t> to match; and the
`tour` reference aliases in update()/timingUpdate() are retyped to
keep compiling (their surrounding int locals are Phase 2's target, not
this one).

Verified: full build and test suite pass; ASan/UBSan clean across all
five executables against the real dataset (including 20 extra trials
of example.cc specifically checking for behavior changes -- confirmed
the occasional non-zero-penalty outcome some runs produce is pre-existing
stochastic variance from the algorithm's minIterations floor, not a
regression, since it appears in both old and new builds at a similar
~10-15% rate and none of these edits change any actual computed value,
only the type it's held in). -Wsign-conversion sites drop from 49 to 40,
with all 40 remaining sites now confined to update()/timingUpdate() --
exactly Phase 2's scope.

+++ src/djinni/routes.h
@@ -324,11 +324,11 @@ public:
     neighbor.setF(getF());
     neighbor.setP(getP());
     while (0 == firstswitch)
-      firstswitch = static_cast<int>((numCustomers - 1) * _dis(_prng)) + 1;
+      firstswitch = static_cast<uint32_t>((numCustomers - 1) * _dis(_prng)) + 1;
     uint32_t secondswitch = firstswitch;
     while ((secondswitch == firstswitch) || (secondswitch == firstswitch - 1))
-      secondswitch = static_cast<int>((numCustomers - 1) * _dis(_prng)) + 1;
-    int holder = _solution[firstswitch];
+      secondswitch = static_cast<uint32_t>((numCustomers - 1) * _dis(_prng)) + 1;
+    uint32_t holder = _solution[firstswitch];
     if (firstswitch < secondswitch) {
       std::copy(_solution.begin(), _solution.begin() + firstswitch,
                 neighbor._solution.begin());
@@ -382,7 +382,7 @@ protected:
     int firstswitch = _firstswitch;
     int secondswitch = _secondswitch;
     int numCustomers = _solution.size();
-    const std::vector<int> &tour = _solution;
+    const std::vector<uint32_t> &tour = _solution;
     const Matrix<double, 2> &travTime = _w->travelTimes();
     if (firstswitch <= secondswitch) {
       if (secondswitch != (numCustomers - 1)) {
@@ -427,7 +427,7 @@ public:
   /*! Randomize this TravelingSalesmanSolution. */
   void randomize() {
     for (uint32_t i = 0; i < _solution.size(); i += 1)
-      _solution[i] = static_cast<int>(i);
+      _solution[i] = i;
     auto iter = _solution.begin();
     ++iter;
     std::shuffle(iter, _solution.end(), _prng);
@@ -525,7 +525,7 @@ public:
   @param os The output stream to dump our path to
   @return The output stream after we've dumped in it */
   std::ostream &dump(std::ostream &os) const {
-    std::ostream_iterator<int> oiter(os, " ");
+    std::ostream_iterator<uint32_t> oiter(os, " ");
     std::ranges::copy(_solution, oiter);
     return os;
   }
@@ -593,7 +593,7 @@ protected:
     const Matrix<double, 2> &travTime = _w->travelTimes();
     const std::vector<double> &lowdeadlines = _w->lowDeadlines();
     const std::vector<double> &deadlines = _w->deadlines();
-    const std::vector<int> &tour = _solution;
+    const std::vector<uint32_t> &tour = _solution;
 
     if (_firstswitch < _secondswitch)
       start = _firstswitch;
@@ -617,7 +617,7 @@ protected:
   }
 
   std::shared_ptr<WorldType> _w;
-  std::vector<int> _solution;
+  std::vector<uint32_t> _solution;
   double _f, _p;
   std::vector<double> _arrivaltime;
   std::vector<double> _penaltysum;
